feat(app): reorganize math verifier
